订单页面的开发

+++ app/cart/biz/service/add_item.go
@@ -31,7 +31,7 @@ func (s *AddItemService) Run(req *cart.AddItemReq) (resp *cart.AddItemResp, err 
 	cartItem := &model.Cart{
 		UserId:    req.UserId,
 		ProductId: req.Item.ProductId,
-		Qty:       req.Item.Quantity,
+		Qty:       int32(req.Item.Quantity),
 	}
 
 	err = model.AddItem(s.ctx, mysql.DB, cartItem)
